Fix react import where it was missing
