+++ tsconfig.json
@@ -3,7 +3,7 @@
     "checkJs": true,
     "noEmit": true,
     "lib": ["DOM", "ES6", "ES2019"],
-    "jsx": "preserve",
+    "jsx": "react",
     "allowSyntheticDefaultImports": true,
     "isolatedModules": true,
     "strictNullChecks": true,
